% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{calc_act_cpp}
\alias{calc_act_cpp}
\title{Calculate the auto correlation time
from \url{https://github.com/beast-dev/beast-mcmc/blob/800817772033c13061f026226e41128d21fd14f3/src/dr/inference/trace/TraceCorrelation.java#L159} # nolint}
\usage{
calc_act_cpp(sample, sample_interval)
}
\arguments{
\item{sample}{sample}

\item{sample_interval}{sample interval}
}
\value{
the auto correlation time
}
\description{
Calculate the auto correlation time
from \url{https://github.com/beast-dev/beast-mcmc/blob/800817772033c13061f026226e41128d21fd14f3/src/dr/inference/trace/TraceCorrelation.java#L159} # nolint
}
\author{
Richèl J.C. Bilderbeek
}
