% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mode.R
\encoding{UTF-8}
\name{calc_mode}
\alias{calc_mode}
\title{Calculate the mode of values
If the distribution is bi or multimodal or uniform, NA is returned}
\usage{
calc_mode(values)
}
\arguments{
\item{values}{numeric vector to calculate the mode of}
}
\value{
the mode of the trace
}
\description{
Calculate the mode of values
If the distribution is bi or multimodal or uniform, NA is returned
}
\examples{
# In a unimodal distribution, find the value that occurs most
calc_mode(c(1, 2, 2))
calc_mode(c(1, 1, 2))

# For a uniform distribution, NA is returned
tracerer:::calc_mode(c(1, 2))
}
\author{
Richèl J.C. Bilderbeek
}
