% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\encoding{UTF-8}
\name{calc_std_error_of_mean_cpp}
\alias{calc_std_error_of_mean_cpp}
\title{Calculates the standard error of the mean}
\usage{
calc_std_error_of_mean_cpp(sample)
}
\arguments{
\item{sample}{numeric vector of values}
}
\value{
the standard error of the mean
}
\description{
Calculates the standard error of the mean
}
\author{
Richèl J.C. Bilderbeek
}
