% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sum_stats.R
\name{calc_sum_stats}
\alias{calc_sum_stats}
\title{Calculates the Effective Sample Sizes from a parsed BEAST2 log file}
\usage{
calc_sum_stats(traces, sample_interval, burn_in_fraction = 0.1)
}
\arguments{
\item{traces}{a dataframe with traces with removed burn-in}

\item{sample_interval}{the interval in timesteps between samples}

\item{burn_in_fraction}{the fraction that needs to be removed, must be [0,1>}
}
\value{
the effective sample sizes
}
\description{
Calculates the Effective Sample Sizes from a parsed BEAST2 log file
}
\author{
Richel J.C. Bilderbeek
}
