% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\encoding{UTF-8}
\name{default_params_doc}
\alias{default_params_doc}
\title{Documentation of general function arguments}
\usage{
default_params_doc(
  log_filename,
  sample_interval,
  state_filename,
  trace,
  tracelog_filename,
  trees_filename,
  trees_filenames,
  verbose
)
}
\arguments{
\item{log_filename}{deprecated name
of the BEAST2 tracelog \code{.log} output file.
Use \code{tracelog_filename} instead}

\item{sample_interval}{the interval in timesteps between samples}

\item{state_filename}{name of the BEAST2 state \code{.xml.state} output file}

\item{trace}{the values}

\item{tracelog_filename}{name of the BEAST2 tracelog \code{.log} output file,
as can be read using \link{parse_beast_tracelog_file}}

\item{trees_filename}{name of a BEAST2 posterior \code{.trees} file,
as can be read using \link{parse_beast_trees}}

\item{trees_filenames}{the names of one or more a BEAST2
posterior \code{.trees} file.
Each \code{.trees} file can be read using \link{parse_beast_trees}}

\item{verbose}{set to TRUE for more output}
}
\description{
This function does nothing.
It is intended to inherit function argument documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
