% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_trees_posterior.R
\encoding{UTF-8}
\name{is_trees_posterior}
\alias{is_trees_posterior}
\title{Determines if the input is a BEAST2 posterior,
as parsed by parse_beast_trees}
\usage{
is_trees_posterior(x)
}
\arguments{
\item{x}{the input}
}
\value{
TRUE or FALSE
}
\description{
Determines if the input is a BEAST2 posterior,
as parsed by parse_beast_trees
}
\author{
Richèl J.C. Bilderbeek
}
