% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_beast_trees.R
\encoding{UTF-8}
\name{parse_beast_trees}
\alias{parse_beast_trees}
\title{Parses a BEAST2 .trees output file}
\usage{
parse_beast_trees(filename)
}
\arguments{
\item{filename}{name of the BEAST2 .trees output file}
}
\value{
the phylogenies in the posterior
}
\description{
Parses a BEAST2 .trees output file
}
\examples{
trees_filename <- get_tracerer_path("beast2_example_output.trees")
parse_beast_trees(trees_filename)
}
\seealso{
Use \code{\link{save_beast_trees}} to save the phylogenies
  to a \code{.trees} file.
  Use \link{is_trees_file} with \code{verbose = TRUE} to find out
  why a file is invalid
}
\author{
Richèl J.C. Bilderbeek
}
