% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_files.R
\name{remove_files}
\alias{remove_files}
\title{Removes files if and only if present}
\usage{
remove_files(filenames)
}
\arguments{
\item{filenames}{names of the files to be removed}
}
\value{
nothing, removes those files that are present
}
\description{
Removes files if and only if present
}
\examples{
  # Can safely delete absent files
  beautier:::remove_files("abs.ent")

  # Create a file
  fasta_filename <- "remove_files.fas"
  beastier:::create_random_fasta(
    n_taxa = 5,
    sequence_length = 10,
    fasta_filename = fasta_filename
  )
  testit::assert(file.exists(fasta_filename))

  # Can safely delete a mix of present and absent files
  beautier:::remove_files(c("abs.ent", fasta_filename))
  testit::assert(!file.exists(fasta_filename))
}
\author{
Richel J.C. Bilderbeek
}
