% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_beast_estimates.R
\encoding{UTF-8}
\name{save_beast_estimates}
\alias{save_beast_estimates}
\title{Save the BEAST2 estimates as a BEAST2 \code{.log} file
There will be some differences: a BEAST2 \code{.log} file also saves
the model as comments and formats the numbers in a way non-standard to R}
\usage{
save_beast_estimates(estimates, filename)
}
\arguments{
\item{estimates}{a data frame of BEAST2 parameter estimates}

\item{filename}{name of the \code{.log} file to save to}
}
\value{
nothing
}
\description{
Save the BEAST2 estimates as a BEAST2 \code{.log} file
There will be some differences: a BEAST2 \code{.log} file also saves
the model as comments and formats the numbers in a way non-standard to R
}
\seealso{
Use \code{\link{parse_beast_log}} to read a BEAST2 \code{.log} file
}
\author{
Richèl J.C. Bilderbeek
}
