% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_beast_trees.R
\encoding{UTF-8}
\name{save_beast_trees}
\alias{save_beast_trees}
\title{Save the BEAST2 trees as a BEAST2 \code{.log} file
There will be some differences: a BEAST2 \code{.log} file also saves
the model as comments and formats the numbers in a way non-standard to R}
\usage{
save_beast_trees(trees, filename)
}
\arguments{
\item{trees}{BEAST2 posterior trees, of type \code{ape::multiPhylo}}

\item{filename}{name of the \code{.trees} file to save to}
}
\value{
nothing
}
\description{
Save the BEAST2 trees as a BEAST2 \code{.log} file
There will be some differences: a BEAST2 \code{.log} file also saves
the model as comments and formats the numbers in a way non-standard to R
}
\seealso{
Use \code{\link{parse_beast_log}} to read a BEAST2 \code{.log} file
}
\author{
Richèl J.C. Bilderbeek
}
