% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracerer-package.R
\docType{package}
\encoding{UTF-8}
\name{tracerer-package}
\alias{tracerer}
\alias{tracerer-package}
\title{\code{tracerer}: A package to parse BEAST2 output files}
\description{
\code{tracerer} allows to parse BEAST2 input files, using
an R interface. 'tracerer' closely follows the functionality
of Tracer, a GUI tool bundled with BEAST and BEAST2,
including its default settings.
}
\seealso{
These are packages associated with \code{tracerer}:
\itemize{
  \item
    The package \code{beautier} can create
    BEAST2 input files from R

  \item
    The package \code{beastier} can run
    BEAST2 from R

  \item
    The package \code{mauricer} manages
    BEAST2 packages from R

  \item
    The package \code{babette} combines the
    functionality of \code{beautier},
    \code{beastier}, \code{tracerer} and
    \code{mauricer} and
    into a single workflow
}
If something is (still) missing from \code{tracerer},
the \code{coda} package may have the functionality you need.
}
\author{
Richèl J.C. Bilderbeek
}
\keyword{internal}
