% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracerer.R
\docType{package}
\name{tracerer}
\alias{tracerer}
\alias{tracerer-package}
\title{tracerer: A package to parse BEAST2 output files.}
\description{
'tracerer' allows to parse BEAST2 input files, using
an R interface. 'tracerer' closely follows the functionality
of Tracer, a GUI tool bundled with BEAST and BEAST2,
including its default settings.
}
\seealso{
These are packages association with 'tracerer':
\itemize{
  \item{
    The package \link[beautier]{beautier} can create
    BEAST2 input files from R
  }
  \item{
    The package \link[beastier]{beastier} can run
    BEAST2 from R
  }
  \item{
    The package \link[babette]{babette} combines the
    functionality of \link[beautier]{beautier},
    \link[beastier]{beastier} and \link[tracerer]{tracerer}
    into a single workflow
  }
}
}
