% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_geom_mean.R
\name{calc_geom_mean}
\alias{calc_geom_mean}
\title{Calculate the geometric mean}
\usage{
calc_geom_mean(values)
}
\arguments{
\item{values}{a numeric vector of values}
}
\description{
Calculate the geometric mean
}
\author{
Richel J.C. Bilderbeek
}
