% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hpd_interval.R
\name{calc_hpd_interval}
\alias{calc_hpd_interval}
\title{Calculate the Highest Probability Density}
\usage{
calc_hpd_interval(trace, proportion = 0.95)
}
\arguments{
\item{trace}{a numeric vector of values}

\item{propertion}{the proportion of numbers within the interval.
For example, use 0.95 for a 95% interval}
}
\description{
Calculate the Highest Probability Density
}
\author{
Richel J.C. Bilderbeek, after code by J. Heled
From https://github.com/beast-dev/beast-mcmc/blob/98705c59db65e4f406a420bbade949aeecfe05d0/src/dr/stats/DiscreteStatistics.java#L317
}
