% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hpd_interval.R
\name{calc_hpd_interval}
\alias{calc_hpd_interval}
\title{Calculate the Highest Probability Density}
\usage{
calc_hpd_interval(trace, proportion = 0.95)
}
\arguments{
\item{trace}{a numeric vector of values}

\item{proportion}{the proportion of numbers within the interval.
For example, use 0.95 for a 95 percentage interval}
}
\value{
a numeric vector, with at index 1 the lower boundary of the
  interval, and at index 2 the upper boundary of the interval
}
\description{
Calculate the Highest Probability Density
}
\seealso{
Java code can be found here: \url{https://github.com/beast-dev/beast-mcmc/blob/98705c59db65e4f406a420bbade949aeecfe05d0/src/dr/stats/DiscreteStatistics.java#L317} # nolint URLs can be long
}
\author{
The original Java version of the algorithm was from J. Heled,
  ported to R and adapted by Richel J.C. Bilderbeek
}
