% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stderr_mean.R
\name{calc_stderr_mean}
\alias{calc_stderr_mean}
\title{Calculate the standard error of the mean}
\usage{
calc_stderr_mean(trace)
}
\arguments{
\item{trace}{the values}
}
\value{
the standard error of the mean
}
\description{
Calculate the standard error of the mean
}
\examples{
  trace <- sin(seq(from = 0.0, to = 2.0 * pi, length.out = 100))
  stderr_mean <- calc_stderr_mean(trace)
  testthat::expect_equal(stderr_mean, expected = 0.4347425, tolerance = 0.01)
}
\seealso{
Java code can be found here: \url{https://github.com/beast-dev/beast-mcmc/blob/800817772033c13061f026226e41128d21fd14f3/src/dr/inference/trace/TraceCorrelation.java#L159} # nolint URLs can be long
}
\author{
The original Java version of the algorithm was from Remco Bouckaert,
  ported to R and adapted by Richèl J.C. Bilderbeek
}
