% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summary_stats.R
\name{calc_summary_stats}
\alias{calc_summary_stats}
\title{Calculates the Effective Sample Sizes from a parsed BEAST2 log file}
\usage{
calc_summary_stats(trace, sample_interval)
}
\arguments{
\item{trace}{a numeric vector of values. Assumes the burn-in
is removed.}

\item{sample_interval}{the interval in timesteps between samples}
}
\value{
the effective sample sizes
}
\description{
Calculates the Effective Sample Sizes from a parsed BEAST2 log file
}
\seealso{
Use \code{\link{remove_burn_in}} to remove the burn-in
  of a trace
}
\author{
Richel J.C. Bilderbeek
}
