% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_trace.R
\name{check_trace}
\alias{check_trace}
\title{Check if the trace is a valid. Will \link{stop} if not}
\usage{
check_trace(trace)
}
\arguments{
\item{trace}{the values}
}
\description{
Check if the trace is a valid. Will \link{stop} if not
}
\examples{
check_trace(seq(1, 2))
}
\author{
Richèl J.C. Bilderbeek
}
