% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_trees_in_file.R
\name{count_trees_in_file}
\alias{count_trees_in_file}
\title{Count the number of trees in a \code{.trees} file}
\usage{
count_trees_in_file(trees_filename)
}
\arguments{
\item{trees_filename}{Name of a BEAST2 posterior \code{.trees} file.}
}
\description{
Count the number of trees in a \code{.trees} file
}
\author{
Richèl J.C. Bilderbeek
}
