% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_trees_in_file.R
\name{count_trees_in_file}
\alias{count_trees_in_file}
\title{Count the number of trees in a \code{.trees} file}
\usage{
count_trees_in_file(trees_filename)
}
\arguments{
\item{trees_filename}{name of a BEAST2 posterior \code{.trees} file,
as can be read using \link{parse_beast_trees}}
}
\value{
the number of trees
}
\description{
Count the number of trees in a \code{.trees} file
}
\seealso{
if the \code{.trees} file is invalid,
  use \link{is_trees_file} with \code{verbose = TRUE}
  for the reason
}
\author{
Richèl J.C. Bilderbeek
}
