% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_std_dev.R
\name{cs_std_dev}
\alias{cs_std_dev}
\title{Corrected Sample standard deviation}
\usage{
cs_std_dev(values)
}
\arguments{
\item{values}{numeric values}
}
\description{
Corrected Sample standard deviation
}
\author{
Richel J.C. Bilderbeek
}
