% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_std_dev.R
\name{cs_std_dev}
\alias{cs_std_dev}
\title{Calculate the corrected sample standard deviation.}
\usage{
cs_std_dev(values)
}
\arguments{
\item{values}{numeric values}
}
\value{
the corrected sample standard deviation
}
\description{
Calculate the corrected sample standard deviation.
}
\author{
Richèl J.C. Bilderbeek
}
