% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_param_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(bd_tree_prior, cbs_tree_prior, ccp_tree_prior,
  cep_tree_prior, clock_model, clock_models, crown_age, crown_ages, distr_id,
  fasta_filename, fasta_filenames, fixed_crown_age, fixed_crown_ages,
  gtr_site_model, hky_site_model, id, ids, initial_phylogenies,
  input_fasta_filenames, is_first, mcmc, misc_options, output_xml_filename,
  param_id, rln_clock_model, sequence_length, site_model, site_models,
  strict_clock_model, tn93_site_model, tree_prior, tree_priors, verbose,
  yule_tree_prior)
}
\arguments{
\item{bd_tree_prior}{a Birth-Death tree prior, as created
by \code{\link{create_bd_tree_prior}}}

\item{cbs_tree_prior}{a Coalescent Bayesian Skyline tree prior,
as returned by \code{\link{create_cbs_tree_prior}}}

\item{ccp_tree_prior}{a Coalescent Constant Population tree prior,
as returned by \code{\link{create_ccp_tree_prior}}}

\item{cep_tree_prior}{a cep tree prior,
as returned by \code{\link{create_cep_tree_prior}}}

\item{clock_model}{a clock model,
as returned by \code{\link{create_clock_model}}}

\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}

\item{crown_age}{the crown age of the phylogeny}

\item{crown_ages}{the crown ages of the phylogenies. Set to NA
if the crown age needs to be estimated}

\item{distr_id}{a distributions' ID}

\item{fasta_filename}{a FASTA filename.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}

\item{fasta_filenames}{One or more FASTA filenames.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}

\item{fixed_crown_age}{determines if the phylogeny's crown age is
fixed. If FALSE, crown age is estimated by BEAST2. If TRUE,
the crown age is fixed to the crown age
of the initial phylogeny.}

\item{fixed_crown_ages}{one or more booleans to determine if the
phylogenies' crown ages are fixed.
If FALSE, crown age is estimated by BEAST2. If TRUE,
the crown age is fixed to the crown age
of the initial phylogeny.}

\item{gtr_site_model}{a GTR site model,
as returned by \code{\link{create_gtr_site_model}}}

\item{hky_site_model}{an HKY site model,
as returned by \code{\link{create_hky_site_model}}}

\item{id}{an alignment's IDs.
An ID can be extracted from its FASTA filesname
with \code{\link{get_ids}})}

\item{ids}{one or more alignments' IDs.
IDs can be extracted from their FASTA filesnames
with \code{\link{get_ids}})}

\item{initial_phylogenies}{one or more MCMC chain's initial phylogenies.
Each one set to NA will result in BEAST2 using a random phylogeny. Else
the phylogeny is assumed to be of class ape::phylo.}

\item{input_fasta_filenames}{One or more FASTA filenames.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}

\item{is_first}{boolean to indicate if this is the first
out of many (one, two or more)}

\item{mcmc}{one mcmc object,
as returned by \code{\link{create_mcmc}}}

\item{misc_options}{one misc_options object,
as returned by \code{\link{create_misc_options}}}

\item{output_xml_filename}{Name of the XML parameter file created by this
function. BEAST2 uses this file as input.}

\item{param_id}{a parameter's ID}

\item{rln_clock_model}{a Relaxed Log-Normal clock model,
as returned by \code{\link{create_rln_clock_model}}}

\item{sequence_length}{a DNA sequence length, in base pairs}

\item{site_model}{a site model,
as returned by \code{\link{create_site_model}}}

\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}

\item{strict_clock_model}{a strict clock model,
as returned by \code{\link{create_strict_clock_model}}}

\item{tn93_site_model}{a TN93 site model,
as returned by \code{\link{create_tn93_site_model}}}

\item{tree_prior}{a tree priors,
as returned by \code{\link{create_tree_prior}}}

\item{tree_priors}{one or more tree priors,
as returned by \code{\link{create_tree_prior}}}

\item{verbose}{if TRUE, additional information is displayed, that
is potentially useful in debugging}

\item{yule_tree_prior}{a Yule tree_prior,
as created by \code{\link{create_yule_tree_prior}}}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\author{
Richel J.C. Bilderbeek
}
