% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path.R
\name{get_tracerer_path}
\alias{get_tracerer_path}
\title{Get the full path of a file in the \code{inst/extdata} folder}
\usage{
get_tracerer_path(filename)
}
\arguments{
\item{filename}{the file's name, without the path}
}
\value{
the full path to the filename
}
\description{
Get the full path of a file in the \code{inst/extdata} folder
}
\examples{
get_tracerer_path("beast2_example_output.log")
get_tracerer_path("beast2_example_output.trees")
get_tracerer_path("beast2_example_output.xml")
get_tracerer_path("beast2_example_output.xml.state")
}
\seealso{
for more files, use \code{\link{get_tracerer_paths}}
}
\author{
Richèl J.C. Bilderbeek
}
