% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_posterior.R
\name{is_posterior}
\alias{is_posterior}
\title{Determines if the input is a BEAST2 posterior}
\usage{
is_posterior(x)
}
\arguments{
\item{x}{the input}
}
\value{
TRUE or FALSE
}
\description{
Determines if the input is a BEAST2 posterior
}
\examples{
  trees_filename <- system.file(
   "extdata", "beast2_example_output.trees", package = "beastier"
  )
  log_filename <- system.file(
   "extdata", "beast2_example_output.log", package = "beastier"
  )
  posterior <- parse_beast_posterior(
    trees_filename = trees_filename,
    log_filename = log_filename
  )
  testit::assert(is_posterior(posterior))
}
\author{
Richel Bilderbeek
}
