% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_posterior.R
\name{is_posterior}
\alias{is_posterior}
\title{Determines if the input is a BEAST2 posterior}
\usage{
is_posterior(x)
}
\arguments{
\item{x}{the input}
}
\value{
TRUE if the input contains all information of
  a BEAST2 posterior. Returns FALSE otherwise.
}
\description{
Determines if the input is a BEAST2 posterior
}
\examples{
  trees_filename <- get_tracerer_path("beast2_example_output.trees")
  log_filename <- get_tracerer_path("beast2_example_output.log")
  posterior <- parse_beast_posterior(
    trees_filename = trees_filename,
    log_filename = log_filename
  )
  testit::assert(is_posterior(posterior))
}
\author{
Richèl J.C. Bilderbeek
}
