% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_trees_file.R
\name{is_trees_file}
\alias{is_trees_file}
\title{Measure if a file a valid BEAST2 \code{.trees} file}
\usage{
is_trees_file(trees_filename, verbose = FALSE)
}
\arguments{
\item{trees_filename}{Name of a BEAST2 posterior \code{.trees} file.}
}
\value{
the phylogenies in the posterior
}
\description{
Measure if a file a valid BEAST2 \code{.trees} file
}
\examples{
  trees_filename <- get_tracerer_path("beast2_example_output.trees")
  posterior <- parse_beast_trees(trees_filename)
  testit::assert(is_trees_posterior(posterior))
}
\seealso{
Use \code{\link{save_beast_trees}} to save the phylogenies
  to a \code{.trees} file.
}
\author{
Richèl J.C. Bilderbeek
}
