% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_beast_log.R
\name{parse_beast_log}
\alias{parse_beast_log}
\title{Deprecated function to parse a BEAST2 \code{.log} output file.
Use \link{parse_beast_tracelog_file} instead}
\usage{
parse_beast_log(tracelog_filename, filename = "deprecated")
}
\arguments{
\item{tracelog_filename}{name of the BEAST2 tracelog \code{.log} output file,
as can be read using \link{parse_beast_tracelog_file}}

\item{filename}{deprecated name of the BEAST2 .log output file}
}
\value{
data frame with the parameter estimates
}
\description{
Deprecated function to parse a BEAST2 \code{.log} output file.
Use \link{parse_beast_tracelog_file} instead
}
\examples{
# Deprecated
parse_beast_log(
  tracelog_filename = get_tracerer_path("beast2_example_output.log")
)
# Use the function 'parse_beast_tracelog_file' instead
parse_beast_tracelog_file(
  tracelog_filename = get_tracerer_path("beast2_example_output.log")
)
}
\author{
Richèl J.C. Bilderbeek
}
