% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_beast_output_files.R
\name{parse_beast_output_files}
\alias{parse_beast_output_files}
\title{Parse all BEAST2 output files}
\usage{
parse_beast_output_files(log_filename, trees_filenames, state_filename,
  burn_in_fraction = 0.1)
}
\arguments{
\item{log_filename}{name of the BEAST2 .log file}

\item{trees_filenames}{name(s) of the BEAST2 .trees file(s) created.
BEAST2 will create one .trees file per alignment}

\item{state_filename}{name of the BEAST2 .xml.state file created}

\item{burn_in_fraction}{the fraction of the parameter estimate that needs
to be removed, must be [0,1>}
}
\description{
Parse all BEAST2 output files
}
\examples{
  trees_filenames <- get_path("beast2_example_output.trees")
  log_filename <- get_path("beast2_example_output.log")
  state_filename <- get_path("beast2_example_output.xml.state")
  out <- parse_beast_output_files(
    log_filename = log_filename,
    trees_filenames = trees_filenames,
    state_filename = state_filename
  )
  testit::assert("estimates" \%in\% names(out))
  testit::assert("beast2_example_output_trees" \%in\% names(out))
  testit::assert("operators" \%in\% names(out))
}
\author{
Richel J.C. Bilderbeek
}
