% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_beast_posterior.R
\name{parse_beast_posterior}
\alias{parse_beast_posterior}
\title{Parses BEAST2 output files to a posterior}
\usage{
parse_beast_posterior(trees_filenames, log_filename, burn_in_fraction = 0.1)
}
\arguments{
\item{trees_filenames}{one or more names of the BEAST2 .trees output files.
A BEAST2 run will create as much .trees files as there are alignments}

\item{log_filename}{name of the BEAST2 .trees output file}

\item{burn_in_fraction}{the fraction of the parameter estimate that needs
to be removed, must be [0,1>}
}
\value{
a posterior
}
\description{
Parses BEAST2 output files to a posterior
}
\examples{
  trees_filenames <- get_path("beast2_example_output.trees")
  log_filename <- get_path("beast2_example_output.log")
  posterior <- parse_beast_posterior(
    trees_filenames = trees_filenames,
    log_filename = log_filename
  )
  testit::assert(is_posterior(posterior))
}
\author{
Richel J.C. Bilderbeek
}
