% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_beast_state.R
\name{parse_beast_state_operators}
\alias{parse_beast_state_operators}
\title{Parses a BEAST2 state \code{.xml.state} output file to get only the operators
  acceptances}
\usage{
parse_beast_state_operators(
  state_filename = get_tracerer_path("beast2_example_output.xml.state"),
  filename = "deprecated"
)
}
\arguments{
\item{state_filename}{name of the BEAST2 state \code{.xml.state} output file}

\item{filename}{deprecated name of the BEAST2 .xml.state output file,
use \code{state_filename} instead}
}
\value{
data frame with all the operators' success rates
}
\description{
Parses a BEAST2 state \code{.xml.state} output file to get only the operators
  acceptances
}
\examples{
parse_beast_state_operators(
  state_filename = get_tracerer_path("beast2_example_output.xml.state")
)
}
\author{
Richèl J.C. Bilderbeek
}
