% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_beast_tracelog_file.R
\name{parse_beast_tracelog_file}
\alias{parse_beast_tracelog_file}
\title{Parses a BEAST2 tracelog \code{.log} output file}
\usage{
parse_beast_tracelog_file(tracelog_filename)
}
\arguments{
\item{tracelog_filename}{name of the BEAST2 tracelog \code{.log} output file,
as can be read using \link{parse_beast_tracelog_file}}
}
\value{
data frame with the parameter estimates
}
\description{
Parses a BEAST2 tracelog \code{.log} output file
}
\examples{
parse_beast_tracelog_file(
  tracelog_filename = get_tracerer_path("beast2_example_output.log")
)
}
\seealso{
Use \code{\link{remove_burn_ins}} to remove the burn-in from
  the returned parameter estimates.
  Use \code{\link{save_beast_estimates}} to save the estimates
  to a \code{.log} file.
}
\author{
Richèl J.C. Bilderbeek
}
