% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_beast_trees.R
\name{parse_beast_trees}
\alias{parse_beast_trees}
\title{Parses a BEAST2 .trees output file}
\usage{
parse_beast_trees(filename, return_type = "list")
}
\arguments{
\item{filename}{name of the BEAST2 .trees output file}

\item{return_type}{the data type of the returned value. Can be either
'list' or 'multiPhylo'}
}
\value{
the phylogenies in the posterior
}
\description{
Parses a BEAST2 .trees output file
}
\examples{
  trees_filename <- get_path("beast2_example_output.trees")
  posterior <- parse_beast_trees(trees_filename)
  testit::assert(is_trees_posterior(posterior))
}
\author{
Richel J.C. Bilderbeek
}
