% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracerer.R
\docType{package}
\name{tracerer}
\alias{tracerer}
\alias{tracerer-package}
\title{tracerer: A package to parse BEAST2 output files.}
\description{
'tracerer' allows to parse BEAST2 input files, using
an R interface. 'tracerer' closely follows the functionality
of Tracer, a GUI tool bundled with BEAST and BEAST2,
including its default settings.
}
\seealso{
These are packages associated with 'tracerer':
\itemize{
  \item{
    The package 'beautier' can create
    BEAST2 input files from R
  }
  \item{
    The package 'beastier' can run
    BEAST2 from R
  }
  \item{
    The package 'babette' combines the
    functionality of 'beautier',
    'beastier' and 'tracerer'
    into a single workflow
  }
}
If something is (still) missing from \code{tracerer},
the \code{coda} package may have the functionality you need.
}
